BE/Feat/#172 : qr코드 추가 및 토큰 수정

See merge request s12-blockchain-nft-sub1/S12P21D210!235

+++ Backend/src/main/java/ssafy/d210/backend/dto/request/payment/TokenRequest.java
@@ -3,9 +3,11 @@ package ssafy.d210.backend.dto.request.payment;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 public class TokenRequest {
     long userId;
-    int quantity;
+    BigInteger quantity;
 }
